e1000-emu: use generalized epoll_callback
